pyproject.toml
[tool.poetry]
name = "e-commerce-platform-app"
version = "0.1.0"
description = "e-commerce_platform_app"
authors = ["nazneenprojects <mulanisnaaz@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "app" },
    { include = "alembic" }
]

[tool.poetry.dependencies]
python = ">=3.13"
sqlalchemy = ">=2.0.36,<3.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
alembic = ">=1.14.0,<2.0.0"
structlog = ">=24.4.0,<25.0.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pyment = "^0.3.3"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"
